autograder buddy to long

+++ proj3/src/core/AutograderBuddy.java
@@ -21,7 +21,7 @@ public class AutograderBuddy {
 //        if (input.contains(":q")) {
 //            return new TETile[][];
 //        }
-        int newSeed = Integer.parseInt(input.substring(1, input.length() - 1));
+        long newSeed = Long.parseLong(input.substring(1, input.length() - 1));
         World gameStart = new World(newSeed);
         return gameStart.returnWorld();
